pyproject.toml
[tool.poetry]
name = "nonebot-plugin-manageweb"
version = "0.1.4"
description = "Nonebot插件商店安装插件，使用基于fastapi和amis的webui安装"
authors = ["CM-Edelweiss <2633668472@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
nonebot2 = "^2.2.0"
fastapi =  ">=0.93.0"
importlib-metadata = ">=4.13.0"
httpx = "^0.23.0"
amis-python = "^1.0.6"
python-jose = "^3.3.0"
toml = ">=0.10.2"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
